move a few functions from utils/index to utils/dates - new fn validDate()
